refactor(label): convert files to typescript
