--- src/__internal__/label/index.d.ts
@@ -1,1 +1,0 @@
-export { default } from "./label";
