--- src/__internal__/label/index.js
@@ -1,1 +1,0 @@
-export { default } from "./label.component";
