+++ src/__internal__/label/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./label.component";
+export type { LabelProps } from "./label.component";
